Fixed serialization of arrays
`extend true` adds properties added to the `Array` prototype to the extended object and they appear in the querystring of the search URL as parameters. Non-deep extends fix that because the first level is always the properties of a plain object, but other levels are not cloned, original references are used. So be warned.

Signed-off-by: Carlos Escribano <carlos@doofinder.com>