Manu1080/azkaban 1753/update library doophoenix v6 Part 2 (#313)
Fixes: doofinder/doophoenix#1753
Part 1: #312

Summary
The problem arose during the hackathon of updating our JavaScript library that works with Doophenix V5 to version 6. Together with @davidmolinacano, we reviewed the new Doophenix API and managed to make all the endpoints work, but there is still a lot of work to be done. We created this issue to finish it.

This is the second in a series of several pull requests that work on the stats module and client.

Content
Add new HTTP methods
Update stats class. (Remove deprecated attributes and add new ones)
Update client. (Remove deprecated methods and add new ones.)
Fix unit tests
Update docs
New version 7