Update library doophoenix v6 part3 (#314)
Fixes: doofinder/doophoenix#1753
Part 1: #312
Part 2: #312

Summary
The problem arose during the hackathon of updating our JavaScript library that works with Doophenix V5 to version 6. Together with @davidmolinacano, we reviewed the new Doophenix API and managed to make all the endpoints work, but there is still a lot of work to be done. We created this issue to finish it.

Update example files.